l commit mta3 ekher nhar

+++ app/src/main/java/com/example/gestiondecommerce/interface_admin.java
@@ -16,13 +16,14 @@ import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
-import com.google.firebase. firestore.CollectionReference;
+import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
@@ -50,58 +51,43 @@ public class interface_admin extends AppCompatActivity {
     Spinner spinner;
     RecyclerView recyclerView;
     private Button btnDatePicker;
+    private Button btnDatePicker1;
     private Calendar selectedDate;
+    private Calendar selectedDate1;
     private String date;
+    private String date1;
     Button chercher;
+    Button efface ;
+    private List<MVT> filteredList;
+
+
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_interface_admin);
         mvtList = new ArrayList<>();
-        spinner =  findViewById(R.id.spinner);
-        chercher = findViewById(R.id.cher);
+        spinner = findViewById(R.id.spinner);
         btnDatePicker = findViewById(R.id.btnDatePicker);
-        btnDatePicker.setOnClickListener(view -> showDatePickerDialog());
+        btnDatePicker1 = findViewById(R.id.btnDatePicker1);
+        efface = findViewById(R.id.effacer);
+
+        btnDatePicker.setOnClickListener(view -> showDatePickerDialog(btnDatePicker));
+        btnDatePicker1.setOnClickListener(view -> showDatePickerDialog(btnDatePicker1));
+        efface.setOnClickListener(view -> clean1());
 
         getUsers(users -> {
-            ArrayAdapter<User> adapter = new ArrayAdapter<>(interface_admin.this, android.R.layout.simple_spinner_item,users);
+            ArrayAdapter<User> adapter = new ArrayAdapter<>(interface_admin.this, android.R.layout.simple_spinner_item, users);
             adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
             spinner.setAdapter(adapter);
         });
 
         recyclerView = findViewById(R.id.rv);
 
-        Button btn = findViewById(R.id.button);
-        btn.setOnClickListener(view -> updateValidationCAdmin());
-
-
-        chercher.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                loadDataFromFirestore(mvtList -> {
-                    commercialAdapter = new CommercialAdapter(mvtList, interface_admin.this);
-                    recyclerView.setLayoutManager(new LinearLayoutManager(interface_admin.this));
-                    recyclerView.setAdapter(commercialAdapter);
-                    recyclerView.setVisibility(View.VISIBLE);
-                });
-            }
-        });
-
         androidx.appcompat.widget.Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
 
         getSupportActionBar().setTitle("Historique");
-
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-
-        Button efface = findViewById(R.id.effacer);
-        efface.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                clean();
-
-            }
-        });
     }
 
     @Override
@@ -115,13 +101,12 @@ public class interface_admin extends AppCompatActivity {
         }
     }
 
-    public void loadDataFromFirestore( MvtsCallBack callBack ) {
+    private void loadDataFromFirestore() {
         User user = (User) spinner.getSelectedItem();
-        if(date != null && user != null) {
+        if (user != null) {
             mvtCollection
                     .whereEqualTo("commercial", user.getName())
-                    .whereEqualTo("validation_admin", false)
-                    .whereEqualTo("date", date)
+                    .whereEqualTo("validation_admin", true)
                     .get()
                     .addOnCompleteListener(task -> {
                         if (task.isSuccessful()) {
@@ -130,56 +115,50 @@ public class interface_admin extends AppCompatActivity {
                                 MVT mvt = document.toObject(MVT.class);
                                 mvt.setId(document.getId());
                                 mvtList.add(mvt);
+                                Log.d("FirestoreData", "Date: " + mvt.getDate());
                             }
-                            if(mvtList.isEmpty()) Toast.makeText(this,"Liste vide",Toast.LENGTH_SHORT).show();
-                            else Toast.makeText(this,"List Non vide",Toast.LENGTH_SHORT).show();
-                            calculateAndSetTotal();
-                            callBack.onCallBack(mvtList);
+                            if (mvtList.isEmpty()) {
+                                Toast.makeText(this, "Liste vide", Toast.LENGTH_SHORT).show();
+                            } else {
+                                filterListByDate(); // Filtrer la liste par date
+                            }
                         } else {
-                            Toast.makeText(this, "Erreur lors du chargement des données Firestore", Toast.LENGTH_SHORT).show();
+                            Log.d("erreur","erreur");
                         }
                     });
-        }else{
-            Toast.makeText(this,"La liste est vide", Toast.LENGTH_SHORT).show();
+        } else {
+            Toast.makeText(this, "La liste est vide", Toast.LENGTH_SHORT).show();
         }
     }
 
-    private void updateValidationCAdmin() {
-        if (mvtList != null && !mvtList.isEmpty()) {
+    private void filterListByDate() {
+        if (date != null && date1 != null) {
+            filteredList = new ArrayList<>();
             for (MVT mvt : mvtList) {
-                String idClient = mvt.getIdClient();
-                if (idClient != null && !mvt.isValidation_admin()) {
-                    mvtCollection
-                            .get()
-                            .addOnSuccessListener(queryDocumentSnapshots -> {
-                                for (QueryDocumentSnapshot document : queryDocumentSnapshots) {
-                                    mvtCollection
-                                            .document(document.getId())
-                                            .update("validation_admin", true)
-                                            .addOnSuccessListener(aVoid -> {
-                                                mvt.setValidation_admin(true);
-                                                loadDataFromFirestore(mvtList -> {
-                                                    commercialAdapter = new CommercialAdapter(mvtList, interface_admin.this);
-                                                    recyclerView.setLayoutManager(new LinearLayoutManager(interface_admin.this));
-                                                    recyclerView.setAdapter(commercialAdapter);
-                                                });
-                                            })
-                                            .addOnFailureListener(e -> {
-                                                Log.e("ValidationError", "Erreur lors de la validation commerciale", e);
-                                                Toast.makeText(this, "Erreur lors de la validation commerciale", Toast.LENGTH_SHORT).show();
-                                            });
-                                }
-                            })
-                            .addOnFailureListener(e -> {
-                                Log.e("ValidationError", "Erreur lors de la récupération des documents", e);
-                                Toast.makeText(this, "Erreur lors de la récupération des documents", Toast.LENGTH_SHORT).show();
-                            });
+                if (mvt.getDate().compareTo(date1) >= 0 && mvt.getDate().compareTo(date) <= 0) {
+                    filteredList.add(mvt);
                 }
             }
-        } else {
-            Log.e("ValidationError", "Liste mvtList est null ou vide");
-            Toast.makeText(this, "Liste mvtList est null ou vide", Toast.LENGTH_SHORT).show();
+            if (filteredList.isEmpty()) {
+                Toast.makeText(this, "Aucune donnée disponible pour la période sélectionnée", Toast.LENGTH_SHORT).show();
+            } else {
+                // Afficher la liste filtrée dans le RecyclerView
+                commercialAdapter = new CommercialAdapter(filteredList, interface_admin.this);
+                recyclerView.setLayoutManager(new LinearLayoutManager(interface_admin.this));
+                recyclerView.setAdapter(commercialAdapter);
+                recyclerView.setVisibility(View.VISIBLE);
+                calculateAndSetTotal(filteredList);
+            }
         }
+    }
+
+    private void calculateAndSetTotal(List<MVT> list) {
+        int sum = 0;
+        for (MVT value : list) {
+            sum += value.getMontant();
+        }
+        TextView t = findViewById(R.id.textView9);
+        t.setText("Total: " + String.valueOf(sum));
     }
 
     private void getUsers(UsersCallback callback){
@@ -188,27 +167,18 @@ public class interface_admin extends AppCompatActivity {
                 .whereEqualTo("role", "commercial")
                 .get()
                 .addOnCompleteListener(task -> {
-                   if(task.isSuccessful()){
-                       for(QueryDocumentSnapshot document :task.getResult()){
-                           User u = document.toObject(User.class);
-                           u.setId(document.getId());
-                           lst.add(u);
-                       }
-                       callback.onCallback(lst);
-                   }
+                    if(task.isSuccessful()){
+                        for(QueryDocumentSnapshot document :task.getResult()){
+                            User u = document.toObject(User.class);
+                            u.setId(document.getId());
+                            lst.add(u);
+                        }
+                        callback.onCallback(lst);
+                    }
                 });
     }
 
-    private void calculateAndSetTotal() {
-        int sum = 0;
-        for (MVT value : mvtList) {
-            sum = sum + value.getMontant();
-        }
-        TextView t = findViewById(R.id.textView9);
-        t.setText("Total: " + String.valueOf(sum));
-    }
-
-    private void showDatePickerDialog() {
+    private void showDatePickerDialog(Button btnDatePicker) {
         final Calendar currentDate = Calendar.getInstance();
         int year = currentDate.get(Calendar.YEAR);
         int month = currentDate.get(Calendar.MONTH);
@@ -217,8 +187,13 @@ public class interface_admin extends AppCompatActivity {
         DatePickerDialog datePickerDialog = new DatePickerDialog(
                 this,
                 (view, yearSelected, monthOfYear, dayOfMonth) -> {
-                    selectedDate = Calendar.getInstance();
+                    Calendar selectedDate = Calendar.getInstance();
                     selectedDate.set(yearSelected, monthOfYear, dayOfMonth);
+                    if (btnDatePicker.getId() == R.id.btnDatePicker) {
+                        this.selectedDate = selectedDate;
+                    } else if (btnDatePicker.getId() == R.id.btnDatePicker1) {
+                        this.selectedDate1 = selectedDate;
+                    }
                     handleDateSelection();
                 },
                 year,
@@ -227,27 +202,39 @@ public class interface_admin extends AppCompatActivity {
         );
         datePickerDialog.show();
     }
+
     private void handleDateSelection() {
-        if (selectedDate != null) {
+        if (selectedDate != null && selectedDate1 != null) {
             SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd", Locale.getDefault());
-            String formattedDate = dateFormat.format(selectedDate.getTime());
-            Toast.makeText(this, "Selected Date: " + formattedDate, Toast.LENGTH_SHORT).show();
-           this.date = formattedDate;
+            date = dateFormat.format(selectedDate.getTime());
+            date1 = dateFormat.format(selectedDate1.getTime());
+            Toast.makeText(this, "Selected Date: " + date, Toast.LENGTH_SHORT).show();
+            Toast.makeText(this, "Selected Date1: " + date1, Toast.LENGTH_SHORT).show();
+            loadDataFromFirestore();
         }
     }
 
-    public void clean(){
-        ProgressDialog pd = new ProgressDialog(this);
-        pd.setMessage("Effacement..");
-        pd.show();
-        for(MVT mvt:mvtList){
-             db.collection("mvt")
-                 .document(mvt.getId())
-                 .delete();
-        }
-        pd.dismiss();
-        Intent intent = new Intent(this,interface_admin.class);
-        startActivity(intent);
-        finish();
+    public void clean1() {
+        AlertDialog.Builder builder = new AlertDialog.Builder(this);
+        builder.setMessage("Êtes-vous sûr de vouloir effacer les éléments sélectionnés ?");
+        builder.setPositiveButton("Oui", (dialog, which) -> {
+            ProgressDialog pd = new ProgressDialog(interface_admin.this);
+            pd.setMessage("Effacement..");
+            pd.show();
+            for (MVT mvt : filteredList) {
+                db.collection("mvt")
+                        .document(mvt.getId())
+                        .delete();
+            }
+            pd.dismiss();
+            Intent intent = new Intent(interface_admin.this, interface_admin.class);
+            startActivity(intent);
+            finish();
+        });
+        builder.setNegativeButton("Non", (dialog, which) -> dialog.dismiss());
+        AlertDialog alertDialog = builder.create();
+        alertDialog.show();
     }
+
+
 }
